piggybank/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
pydantic==2.6.4
pydantic-settings==2.2.1
python-multipart==0.0.9
python-dateutil==2.9.0
